fix: sonar duplicated code and sonar issues
